requirements.txt
torch
accelerate
transformers
diffusers
datasets
huggingface_hub
pandas
matplotlib
scikit-image
Pillow
opencv-python
